requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
urllib3==2.2.1
yarl==1.9.4
